TypeScript Migrate (#28)
* typescript setup

* clean code

* migrate to typescript
