+++ .eslintrc.json
@@ -6,8 +6,10 @@
   "extends": [
     "eslint:recommended",
     "plugin:react/recommended",
-    "plugin:react-hooks/recommended"
+    "plugin:react-hooks/recommended",
+    "plugin:@typescript-eslint/recommended"
   ],
+  "parser": "@typescript-eslint/parser",
   "parserOptions": {
     "ecmaFeatures": {
       "jsx": true
@@ -15,10 +17,8 @@
     "ecmaVersion": "latest",
     "sourceType": "module"
   },
-  "plugins": ["react"],
-  "rules": {
-    "react/prop-types": [0]
-  },
+  "plugins": ["react", "@typescript-eslint"],
+  "rules": {},
   "settings": {
     "react": {
       "version": "detect" // React version. "detect" automatically picks the version you have installed.
