+++ .gitignore
@@ -10,6 +10,7 @@
 
 # production
 /build
+/tsbuild
 
 # misc
 .DS_Store
