+++ src/Components/Logo.tsx
@@ -2,10 +2,11 @@ import React from "react";
 import { Image } from "@chakra-ui/react";
 import logoFile from "../assets/logo192.png";
 
-const Logo = ({boxSize}) => {
-  return (
-    <Image boxSize={boxSize} src={logoFile} alt="Logo" />
-  );
+type Props = {
+  boxSize: string;
+};
+const Logo = ({ boxSize }: Props) => {
+  return <Image boxSize={boxSize} src={logoFile} alt="Logo" />;
 };
 
 export default Logo;
