+++ src/Context/AuthProvider.tsx
@@ -2,7 +2,11 @@ import React, { createContext, useState } from "react";
 
 const AuthContext = createContext({});
 
-export const AuthProvider = ({ children }) => {
+interface Props {
+  children: React.ReactNode;
+}
+
+export const AuthProvider = ({ children }: Props) => {
   const [auth, setAuth] = useState({});
 
   return (
