--- src/Global/toastConfig.js
@@ -1,8 +1,0 @@
-import React from "react";
-
-export const toastConfig = {
-    position: "bottom-left",
-    duration: 4000,
-    isClosable: true,
-    icon: <></>,
-  };
