+++ src/index.tsx
@@ -9,7 +9,9 @@ import { AuthProvider } from "./Context/AuthProvider";
 
 const { ToastContainer } = createStandaloneToast({ theme: theme });
 
-const root = ReactDOM.createRoot(document.getElementById("root"));
+const root = ReactDOM.createRoot(
+  document.getElementById("root") as HTMLElement
+);
 root.render(
   <React.StrictMode>
     <BrowserRouter>
